Add sample get data using mongoDb

+++ buildingBlocks/application/responses/paginationResponse.go
@@ -2,7 +2,7 @@ package responses
 
 type PaginationResponse[TResponse any] struct {
 	Items            []TResponse
-	TotalRecord      int
-	CurrentPageIndex int
-	TotalPage        int
+	TotalRecord      int64
+	CurrentPageIndex int64
+	TotalPage        int64
 }
